+++ src/lib.rs
@@ -7,6 +7,7 @@ use std::fs;
 use std::fs::File;
 use std::io::{self, BufWriter, Write as _};
 use std::path::{Path, PathBuf};
+use std::time::Instant;
 
 use anyhow::Context as _;
 use idalib::IDAError;
@@ -72,6 +73,8 @@ impl ArgHintsMode {
 ///
 /// Returns [`anyhow::Error`] in case something goes wrong with analyzing the binary file or decompiling functions.
 pub fn run(filepath: impl AsRef<Path>) -> anyhow::Result<usize> {
+    let start = Instant::now();
+
     // Open the target binary and run auto-analysis.
     println!(
         "[*] Analyzing binary file `{}`",
@@ -158,8 +161,9 @@ pub fn run(filepath: impl AsRef<Path>) -> anyhow::Result<usize> {
         dirpath.display()
     );
     println!(
-        "[+] Done processing binary file `{}`",
-        filepath.as_ref().display()
+        "[+] Done processing binary file `{}` in {:.1} seconds",
+        filepath.as_ref().display(),
+        start.elapsed().as_secs_f64()
     );
     Ok(decompiled_count)
 }